+++ smart_contracts_for_testing/utils.py
@@ -1,0 +1,5 @@
+
+def sanitise_string(s: str) -> str:
+    """Remove null characters."""
+    # https://stackoverflow.com/a/18762899/315168
+    return s.replace("\x00", "\U0000FFFD")